+++ utils.go
@@ -135,26 +135,31 @@ func deepPassMessages(v Validator, messages Messages) {
 }
 
 // Adds one specific message to passed ruleSet if message is not a empty string
-func addSpecificMessage(r ruleSet, funcName, message string, child bool) {
-	funcName = strings.ReplaceAll(strings.ReplaceAll(gStrings.SnakeCase(funcName), "c_", ""), "child_", "")
+func addSpecificMessage(r ruleSet, funcName, message string) {
+	splits := strings.SplitN(funcName, ".", 2)
+	if firstElementSnake := gStrings.SnakeCase(splits[0]); len(splits) > 1 && (firstElementSnake == "c" || firstElementSnake == "child") {
+		addSpecificMessage(r.getChildrenValidator().getRule(), splits[1], message)
+		return
+	}
+	funcName = gStrings.SnakeCase(funcName)
 	if message != "" {
-		if child {
-			r.getChildrenValidator().getRule().appendSpecificMessages(funcName, message)
-		} else {
-			r.appendSpecificMessages(funcName, message)
-		}
+		r.appendSpecificMessages(funcName, message)
 	}
 }
 
 // Adds rules which are inside passed slice of strings called tag
-func applyRules(r ruleSet, tag []string, o *generatorS, orXor bool) (normalFuncName, funcName string, child bool) {
+func applyRules(r ruleSet, tag []string, o *generatorS, orXor bool) (normalFuncName string) {
 	normalFuncName = strings.TrimSpace(tag[0])
-	if splits := strings.Split(normalFuncName, "."); len(splits) == 2 {
-		funcName = gStrings.PascalCase(splits[0]) + "." + gStrings.PascalCase(splits[1])
-		child = true
+	funcName := normalFuncName
+	splits := strings.SplitN(normalFuncName, ".", 2)
+	if firstElementSnake := gStrings.SnakeCase(splits[0]); len(splits) > 1 && (firstElementSnake == "c" || firstElementSnake == "child") {
+		if r.getChildrenValidator() == nil {
+			r.Children(o.R())
+		}
+		applyRules(r.getChildrenValidator().getRule(), []string{splits[1], tag[1]}, o, orXor)
+		return normalFuncName
 	} else if len(splits) == 1 {
 		funcName = gStrings.PascalCase(normalFuncName)
-		child = false
 	} else {
 		panic(fmt.Sprintf("can't understand %s rule", normalFuncName))
 	}
@@ -261,12 +266,6 @@ func applyRules(r ruleSet, tag []string, o *generatorS, orXor bool) (normalFuncN
 			params = append(params, item)
 		}
 		r.Choices(params...)
-	case "Child.Choices", "C.Choices":
-		params := []interface{}{}
-		for _, item := range parameters {
-			params = append(params, item)
-		}
-		r.Children(o.R().Choices(params...))
 	case "WhenExistOne":
 		r.WhenExistOne(parameters...)
 	case "WhenExistAll":